Add test to verify default threadpool for custom handlers.
+ Move custom handler tests to separate class.
